requirements.txt
torch>=2.1.0
transformers>=4.48.0
accelerate>=0.26.0
nibabel>=5.2.0
numpy>=1.26.0
Pillow>=10.2.0
huggingface_hub>=0.20.0
scipy>=1.11.0
sentencepiece>=0.1.99
protobuf>=4.25.0
matplotlib>=3.8.0
streamlit>=1.30.0
